refactoring

+++ src/controllers/task-controller.js
@@ -4,16 +4,17 @@ class TaskController {
 
   async addTask(req, res) {
     try {
-      const body = req.body;
-      const taskExists = await Task.exists({text: body.text});
-      const newTask = await Task.create({...body, userId: req.userId});
+      if (req.body) {
+        const body = req.body;
+        const taskExists = await Task.exists({text: body.text});
 
-      if (taskExists) {
-        return res.status(403).send({message: 'this task already exists'}
-        )
+        if (!taskExists) {
+          const newTask = await Task.create({...body, userId: req.userId});
+          return res.status(200).send(newTask);
+        } else {
+          return res.status(409).send({message: 'this task already exists'});
+        }
       }
-
-      res.status(200).send(newTask);
     } catch (e) {
       res.status(500).send(`addTask: ${e}`);
     }
@@ -21,14 +22,16 @@ class TaskController {
 
   async editTask(req, res) {
     try {
-      const {_id, text} = req.body;
-      const updatedTask = await Task.updateOne({_id}, req.body);
+      if (req.body) {
+        const {_id, text} = req.body;
 
-      if (text === '') {
-        return res.status(400).send({message: 'text should not be empty'});
+        if (text === '') {
+          return res.status(400).send({message: 'text should not be empty'});
+        }
+
+        const updatedTask = await Task.updateOne({_id}, req.body);
+        res.status(200).send(updatedTask);
       }
-
-      res.status(200).send(updatedTask);
     } catch (e) {
       res.status(500).send(`editTask: ${e}`);
     }
@@ -36,10 +39,12 @@ class TaskController {
 
   async deleteTask(req, res) {
     try {
-      const _id = req.query._id;
-      const result = await Task.deleteOne({_id});
+      if (req.query._id) {
+        const _id = req.query._id;
+        const result = await Task.deleteOne({_id});
 
-      res.status(200).send(result);
+        res.status(200).send(result);
+      }
     } catch (e) {
       res.status(500).send(`deleteTask: ${e}`);
     }
